Move randbytes to synthetic workflow module
