+++ tests/wf/synthetic/utils_test.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 import pytest
 
-from webs.data import randbytes
+from webs.wf.synthetic.utils import randbytes
 
 
 @pytest.mark.parametrize('size', (0, 1, 10, 100))
